Added Itself theme
This theme doesn't really have an overall theme. Each individual logo is based around "itself"! That is to say that all logos are themed around the logo and intent of the program. For example: steampunk for Steam, a Desktop computer in a home office for XFCE, and a blocky world for Prism.